requirements.txt
https://www.djangoproject.com/download/1.7c2/tarball/
django-crispy-forms==1.4.0
django-floppyforms==1.2.0
gnureadline==6.3.3
ipdb==0.8
ipython==2.1.0
wsgiref==0.1.2
